requirements.txt
psycopg2==2.9.10
psycopg2-binary==2.9.10
influxdb-client>=1.36.0
python-dateutil>=2.8.2
